Moved render into separate folder
